+++ model_utils/__init__.py
@@ -1,10 +1,12 @@
-from pkg_resources import DistributionNotFound, get_distribution
-
 from .choices import Choices  # noqa:F401
 from .tracker import FieldTracker, ModelTracker  # noqa:F401
 
 try:
+    from pkg_resources import DistributionNotFound, get_distribution
     __version__ = get_distribution("django-model-utils").version
+except ImportError:  # pragma: no cover
+    # setuptools/pkg-resources is not installed
+    __version__ = None
 except DistributionNotFound:  # pragma: no cover
     # package is not installed
     __version__ = None